Check executingPlayer for hidden turns in DraftViews
This changes the behaviour of hidden turns.

Before, for turns where player and executingPlayer differ, this meant
that the selection was hidden from the player who actually clicked the
panel (and should know the value), and instead shown to the other
player.

Now, we show hidden values exactly to the player who clicked them.